[FEAT] 상단 영역 시간 변경
